fix: refactoring app
