+++ src/entities/category/ui/CategoryModal.tsx
@@ -2,7 +2,7 @@ import { Modal } from 'antd';
 import { Category } from '../types';
 import React from 'react';
 import { CategoryFormValues } from 'src/features/forms/CategoryFrom';
-import { CategoryFormWrapper } from 'src/features/wrappers/CategoryFormWrapper';
+import { CategoryFormWrapper } from 'src/features/wrappers';
 import { useTranslation } from 'react-i18next';
 
 interface CategoryModalProps {
