+++ src/entities/profile/api/index.ts
@@ -1,0 +1,2 @@
+export * from './profileApi';
+export * from './types';
