+++ src/entities/profile/api/types.ts
@@ -1,0 +1,3 @@
+export interface UpdateProfileBody {
+  name: string;
+}
