--- src/entities/profile/model/types.ts
@@ -1,11 +1,0 @@
-export interface Profile {
-  id: string;
-  name: string;
-  email: string;
-  signUpDate: string;
-  commandId: string;
-}
-
-export interface UpdateProfileBody {
-  name: string;
-}
