+++ src/entities/profile/types.ts
@@ -1,4 +1,4 @@
-export interface User {
+export interface Profile {
   id: string;
   email: string;
   name: string;
