+++ src/entities/user/api/index.ts
@@ -1,0 +1,2 @@
+export * from './authApi';
+export * from './types';
