+++ src/entities/user/api/types.ts
@@ -13,14 +13,6 @@ export interface SignInBody {
   password: string;
 }
 
-export interface Profile {
-  id: string;
-  name: string;
-  email: string;
-  signUpDate: string;
-  commandId: string;
-}
-
 export interface UpdateProfileBody {
   name: string;
 }
