--- src/features/forms/ProfileForm/AboutField/AboutField.tsx
@@ -1,46 +1,0 @@
-import React, { memo } from 'react';
-import cn from 'clsx';
-import Input from 'antd/lib/input';
-import type { FormikHandlers } from 'formik/dist/types';
-import { useTranslation } from 'react-i18next';
-import { FormItem } from '../../../../shared/ui-kit/FormItem';
-import { getValidates } from '../../../../utils/validation';
-import { ProfileFormProps } from '../types';
-import s from './AboutField.module.sass';
-
-export type AboutFieldProps = Pick<ProfileFormProps, 'className' | 'disabled'> & {
-  submitCount: number;
-  touched: boolean;
-  errors: string;
-  value: string;
-  onChange: FormikHandlers['handleChange'];
-  onBlur: FormikHandlers['handleBlur'];
-};
-
-export const AboutField = memo(
-  ({ className, onChange, onBlur, touched, value, errors, disabled, submitCount }: AboutFieldProps) => {
-    const { t } = useTranslation();
-
-    const { validateStatus, help } = getValidates(errors, touched, submitCount);
-
-    return (
-      <FormItem
-        className={cn(s.root, className)}
-        title={t(`forms.ProfileForm.about.title`)}
-        validateStatus={validateStatus}
-        help={help}
-      >
-        <Input.TextArea
-          disabled={disabled}
-          name="about"
-          onChange={onChange}
-          onBlur={onBlur}
-          value={value}
-          placeholder={t(`forms.ProfileForm.about.placeholder`)}
-        />
-      </FormItem>
-    );
-  }
-);
-
-AboutField.displayName = 'AboutField';
