--- src/features/forms/ProfileForm/AboutField/index.ts
@@ -1,1 +1,0 @@
-export * from './AboutField';
