+++ src/features/forms/ProfileForm/ProfileForm.stories.tsx
@@ -18,7 +18,6 @@ export const Default: Story = {
     <ProfileFormWrapper
       initialValues={{
         name: '',
-        about: '',
       }}
     />
   ),
