+++ src/features/forms/ProfileForm/types.ts
@@ -2,7 +2,6 @@ import { FormProps } from 'src/features/forms/types';
 
 export type ProfileFormValues = {
   name: string;
-  about: string;
 };
 
 export type ProfileFormErrors = Record<keyof ProfileFormValues, string>;
