+++ src/features/wrappers/ProfileFormWrapper.tsx
@@ -3,23 +3,25 @@ import { Formik } from 'formik';
 import type { FormikHelpers } from 'formik';
 import { ProfileForm } from '../forms/ProfileForm';
 import type { ProfileFormValues } from '../forms/ProfileForm/types';
-import type { InputRef } from 'antd';
+import { Form, type FormInstance, type InputRef } from 'antd';
 import { isNotDefinedString, isTooLongLength } from '../../utils/validation';
 import { useTranslation } from 'react-i18next';
-import { I18nWrapper } from './I18nWrapper';
 
 interface ProfileFormWrapperProps {
   initialValues?: ProfileFormValues;
   disabled?: boolean;
   formRef?: React.MutableRefObject<HTMLFormElement>;
   inputRef?: React.MutableRefObject<InputRef>;
+  onSubmit?: (values: ProfileFormValues, form?: FormInstance) => Promise<void> | void;
 }
 
 export const ProfileFormWrapper: React.FC<ProfileFormWrapperProps> = ({
-  initialValues = { name: '', about: '' },
+  initialValues = { name: '' },
+  onSubmit,
   disabled = false,
 }) => {
   const { t } = useTranslation();
+  const [form] = Form.useForm<{ name: string }>();
 
   const validate = (values: ProfileFormValues) => {
     const errors: Partial<Record<keyof ProfileFormValues, string>> = {};
@@ -30,16 +32,11 @@ export const ProfileFormWrapper: React.FC<ProfileFormWrapperProps> = ({
       errors.name = t(`forms.ProfileForm.name.errors.tooLong`);
     }
 
-    if (isTooLongLength(values.about, 200)) {
-      errors.name = t(`forms.ProfileForm.about.errors.tooLong`);
-    }
-
     return errors;
   };
 
-  const handleSubmit = (values: ProfileFormValues, helpers: FormikHelpers<ProfileFormValues>) => {
-    console.log('ProfileForm submit:', values);
-    helpers.resetForm();
+  const handleSubmit = async (values: ProfileFormValues, helpers: FormikHelpers<ProfileFormValues>) => {
+    await onSubmit?.(values, form);
     helpers.setSubmitting(false);
   };
 
@@ -49,14 +46,7 @@ export const ProfileFormWrapper: React.FC<ProfileFormWrapperProps> = ({
   return (
     <Formik<ProfileFormValues> initialValues={initialValues} validate={validate} onSubmit={handleSubmit}>
       {(formManager) => (
-        <I18nWrapper>
-          <ProfileForm
-            formElement={formRef}
-            formManager={formManager}
-            autoFocusElement={inputRef}
-            disabled={disabled}
-          />
-        </I18nWrapper>
+        <ProfileForm formElement={formRef} formManager={formManager} autoFocusElement={inputRef} disabled={disabled} />
       )}
     </Formik>
   );
