+++ src/features/wrappers/index.ts
@@ -1,5 +1,5 @@
+export * from './CategoryFormWrapper';
 export * from './ChangePasswordFormWrapper';
-export * from './I18nWrapper';
 export * from './OperationFormWrapper';
 export * from './ProfileFormWrapper';
 export * from './SignInFormWrapper';
