+++ src/hooks/useDispatch.ts
@@ -1,0 +1,4 @@
+import { useDispatch } from 'react-redux';
+import { AppDispatch } from 'src/app/store';
+
+export const useAppDispatch = () => useDispatch<AppDispatch>();
